Allows player to give several items at the same time

+++ scripts/actorOverride.js
@@ -27,22 +27,27 @@ function giveItem(currentItemId) {
   const currentActor = this.actor;
   const listPC = game.actors.entities.filter(a => a.hasPlayerOwner);
   const filteredPCList = listPC.filter(a => a.id !== this.actor.id);
-  const d = new PlayerDialog((playerId) => {
+  const d = new PlayerDialog(({playerId, quantity}) => {
       const actor = game.actors.get(playerId);
       const currentItem = currentActor.items.find(item => item.id === currentItemId);
-      const updateItem = {
-        "data.quantity": currentItem.data.data.quantity - 1
-      }
-      currentItem.update(updateItem).then(res => {
-          const duplicatedItem = duplicate(currentItem);
-          duplicatedItem.data.quantity = 1;
-          actor.createEmbeddedEntity("OwnedItem", duplicatedItem);
-          console.log(`Giving item: ${currentItem.id} to actor ${actor.id}`);
-          if (currentItem.data.data.quantity === 0) {
-              currentItem.delete();
-          }
+      const currentItemQuantity = currentItem.data.data.quantity;
+      if (quantity > currentItemQuantity) {
+        return ui.notifications.error(`You cannot offer more items than you posses`);
+      } else {
+        const updateItem = {
+          "data.quantity": currentItem.data.data.quantity - quantity
         }
-      );
+        currentItem.update(updateItem).then(res => {
+            const duplicatedItem = duplicate(currentItem);
+            duplicatedItem.data.quantity = quantity;
+            actor.createEmbeddedEntity("OwnedItem", duplicatedItem);
+            console.log(`Giving item: ${currentItem.id} to actor ${actor.id}`);
+            if (currentItem.data.data.quantity === 0) {
+                currentItem.delete();
+            }
+          }
+        );
+      }
     },
     {acceptLabel: "Offer Item", filteredPCList}
   );
